pyproject.toml
[project]
name = "twitter-capture"
version = "0.1.0"
description = "Telegram bot that converts Twitter/X URLs to images"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "OpenClaw", email = "openclaw@example.com" }
]
keywords = ["twitter", "telegram", "bot", "screenshot", "playwright"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
]

[project.scripts]
twitter-capture = "src.bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
